+++ src/pages/about-us/oaks-curricula.tsx
@@ -9,7 +9,10 @@ import {
 import Layout from "@/components/AppComponents/Layout";
 import { AboutUsLayout } from "@/components/GenericPagesComponents/AboutUsLayout";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
-import { AboutSharedHeader } from "@/components/GenericPagesComponents/AboutSharedHeader";
+import {
+  AboutSharedHeader,
+  AboutSharedHeaderImage,
+} from "@/components/GenericPagesComponents/AboutSharedHeader";
 import { getFeatureFlag } from "@/node-lib/posthog/getFeatureFlag";
 import { getPosthogIdFromCookie } from "@/node-lib/posthog/getPosthogId";
 import getBrowserConfig from "@/browser-lib/getBrowserConfig";
@@ -20,6 +23,7 @@ import SubjectPhasePicker from "@/components/SharedComponents/SubjectPhasePicker
 import { SubjectPhasePickerData } from "@/components/SharedComponents/SubjectPhasePicker/SubjectPhasePicker";
 import { filterValidCurriculumPhaseOptions } from "@/pages-helpers/curriculum/docx/tab-helpers";
 import { CurriculumPartners } from "@/components/GenericPagesComponents/CurriculumPartners";
+import { GuidingPrinciples } from "@/components/GenericPagesComponents/GuidingPrinciples";
 
 const posthogApiKey = getBrowserConfig("posthogApiKey");
 
@@ -52,13 +56,19 @@ export const OaksCurricula: NextPage<OaksCurriculaPage> = ({
           title={"Oak’s curricula"}
           content={pageData.header.textRaw}
           titleHighlight="bg-decorative4-main"
-        />
+        >
+          <AboutSharedHeaderImage
+            imageAlt=""
+            imageUrl="https://res.cloudinary.com/oak-web-application/image/upload/v1734018530/OWA/illustrations/planning-curriculum_xhs7ev.svg"
+          />
+        </AboutSharedHeader>
         <OakBox $background={"bg-decorative4-very-subdued"}>
           <OakMaxWidth $pv={"spacing-80"} $ph={["spacing-16"]}>
             <OakFlex $flexDirection={"column"} $gap={"spacing-56"}>
-              <OakBox $pa={"spacing-16"} $background={"text-inverted"}>
-                TODO: Guiding principals
-              </OakBox>
+              <GuidingPrinciples
+                $background="bg-primary"
+                accentColor="border-decorative4"
+              />
               <OakFlex
                 $flexDirection="column"
                 $maxWidth={"spacing-640"}